% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{defm-names}
\alias{defm-names}
\alias{get_Y_names}
\alias{get_X_names}
\title{Access to the names of a model's datasets}
\usage{
get_Y_names(m)

get_X_names(m)
}
\arguments{
\item{m}{An object of class \link{DEFM}.}
}
\value{
A character vector.
}
\description{
Retrieve the column names of the dependent variable (\code{y}) and independent
variable (\code{x}.)
}
