% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/defm-package.R
\name{DEFM}
\alias{DEFM}
\alias{new_defm_cpp}
\alias{new_defm}
\alias{defm}
\alias{init_defm}
\alias{print_stats}
\alias{nterms_defm}
\alias{nrow_defm}
\alias{ncol_defm_y}
\alias{ncol_defm_x}
\alias{nobs_defm}
\alias{morder_defm}
\title{Discrete Exponential Family Model (DEFM)}
\usage{
new_defm_cpp(id, Y, X, order = 1L)

init_defm(m)

print_stats(m, i = 0L)

nterms_defm(m)

nrow_defm(m)

ncol_defm_y(m)

ncol_defm_x(m)

nobs_defm(m)

morder_defm(m)

new_defm(id, Y, X, order = 1)
}
\arguments{
\item{id}{Integer vector of length \code{n}. Observation ids, for example, person id.}

\item{Y}{0/1 matrix of responses of \code{n_y} columns and \code{n} rows.}

\item{X}{Numeric matrix of covariates of size \code{n_x} by \code{n}.}

\item{order}{Integer. Order of the markov process, by default, 1.}

\item{m}{An object of class \code{DEFM}.}

\item{i}{An integer scalar indicating which set of statistics to print (see details.)}
}
\value{
An external pointer of class \code{DEFM.}
}
\description{
Discrete Exponential Family Models (DEFMs) are models from the exponential
family that deal with discrete data. Here, we deal with binary arrays which
can be used to represent, among other things, networks and multinomial binary
Markov processes.
}
\details{
The \code{print_stats} function prints the supportset of the ith type
of array in the model.
}
