% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defm_mle.R
\name{defm_mle}
\alias{defm_mle}
\title{Maximum Likelihood Estimation of DEFM}
\usage{
defm_mle(object, start, lower, upper, ...)
}
\arguments{
\item{object}{An object of class \link{DEFM}.}

\item{start}{Double vector. Starting point for the MLE.}

\item{lower, upper}{Lower and upper limits for the optimization (passed to
\link[stats4:mle]{stats4::mle}.)}

\item{...}{Further arguments passed to \link[stats4:mle]{stats4::mle}.}
}
\description{
Maximum Likelihood Estimation of DEFM
}
