% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglike_defm}
\alias{loglike_defm}
\title{Log-Likelihood of DEFM}
\usage{
loglike_defm(m, par, as_log = TRUE)
}
\arguments{
\item{m}{An object of class \link{DEFM}}

\item{par}{A vector of parameters of length \code{nterms_defm(m)}.}

\item{as_log}{Logical scalar. When \code{TRUE} (default) returns the log-likelihood,
otherwise it returns the likelihood.}
}
\value{
Numeric, the computed likelihood or log-likelihood of the model.
}
\description{
Log-Likelihood of DEFM
}
