% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_census.R
\name{motif_census}
\alias{motif_census}
\title{Motif census}
\usage{
motif_census(m, locs)
}
\arguments{
\item{m}{An object of class \link{DEFM}.}

\item{locs}{Idx (starting from zero) with the variables that will be
included in the census.}
}
\description{
Motif census
}
