% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defm_mle.R
\name{defm_mle}
\alias{defm_mle}
\alias{summary_table}
\title{Maximum Likelihood Estimation of DEFM}
\usage{
defm_mle(object, start, lower, upper, ...)

summary_table(object, as_texreg = FALSE, ...)
}
\arguments{
\item{object}{An object of class \link{DEFM}.}

\item{start}{Double vector. Starting point for the MLE.}

\item{lower, upper}{Lower and upper limits for the optimization (passed to
\link[stats4:mle]{stats4::mle}.)}

\item{...}{Further arguments passed to \link[stats4:mle]{stats4::mle}.}

\item{as_texreg}{When \code{TRUE}, wraps the result in a texreg object}
}
\description{
Maximum Likelihood Estimation of DEFM
}
\details{
The function \code{summary_table} computes pvalues and returns a table
with the estimates, se, and pvalues. If \code{as_texreg = TRUE}, then it will
return a texreg object.
}
