% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{defm_terms}
\alias{defm_terms}
\alias{term_defm_ones}
\alias{terms_defm}
\alias{term_defm_fe}
\alias{term_defm_transition}
\alias{print.DEFM}
\title{Model specification for DEFM}
\usage{
term_defm_ones(m, idx = -1L)

term_defm_fe(m, idx = -1L, k = 1)

term_defm_transition(m, mat, covar_idx = -1L)

\method{print}{DEFM}(x)
}
\arguments{
\item{m}{An object of class \link{DEFM}.}

\item{idx}{Integer. When greater than -1, index (starting from 0) of a covariate
used to weight the term.}

\item{k}{Numeric scalar. Exponent used in the term.}

\item{mat}{Integer matrix. The matrix specifies the type of motif to capture
(see details.)}
}
\value{
Invisible 0.
}
\description{
Model specification for DEFM
}
\details{
In \code{term_defm_transition}, users can specify a particular motif to model. Motifs
are represented by cells with values equal to 1, for example, the matrix:\if{html}{\out{<div class="sourceCode y0 y1 y2">}}\preformatted{t0:   1 NA NA
t1:   1  1 NA
}\if{html}{\out{</div>}}

represents a transition \verb{y0 -> (y1, y2)}. If 0 is a motif of interest, then
the matrix should include 0 to mark zero values.
}
