% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rule_not_one_to_zero}
\alias{rule_not_one_to_zero}
\title{Add rule for avoiding switching a one to zero in a Markov process}
\usage{
rule_not_one_to_zero(m, idx)
}
\description{
Add rule for avoiding switching a one to zero in a Markov process
}
